Delegated `name` and `description` of objects to `Entity` class
Added an `Entity` class which stores a `name`, a `description`, and a
`name_plural`. `name` no longer set as readable but instead viewed
through a function called `name` which uses the `name_plural` to
pluralise the name of the object if required. This way objects don't
need to handle pluralisation themselves, and the functionality isn't
dumped into a utility function (something I'd like to avoid)